refactor(db): move JSONNullBool to types.go
